responded to feedback in https://github.com/apache/incubator-streams/pull/200
